Add NIC multicast group filter handling and refusal auditing to netstack
- Introduced `MULTICAST_FILTER_REFUSED` event to notify when NIC filter capacity is exceeded.
- Implemented multicast address filter programming in the netstack, including retries on rejection.
- Updated drivers to handle group set programming and provide feedback on refusals.
- Added comprehensive tests for multicast filter behavior, including empty sets, overflows, and retries.
- Updated documentation to reflect the new multicast filter functionality and its implications.

+++ drivers/network/genet/src/regs.rs
@@ -231,7 +231,7 @@ pub const UMAC_MDF_CTRL: usize = UMAC + 0x650;
 pub const UMAC_MDF_ADDR: usize = UMAC + 0x654;
 
 /// Destination-address filter slots the UniMAC provides.
-pub const MDF_SLOTS: u32 = 17;
+pub const MDF_SLOTS: u16 = 17;
 
 /// Bytes one destination-address filter slot occupies.
 pub const MDF_SLOT_STRIDE: usize = 8;
